requirements.txt
fastapi~=0.120.1
SQLAlchemy~=2.0.44
pydantic~=2.12.3
bcrypt~=5.0.0
uvicorn~=0.38.0
PyJWT~=2.10.1
python-dotenv~=1.2.1